Move the bulk of what was in showlogfile, into showlogfile_sup, so that it can be "included" into other scripts; showlogfile, beginexp, swapexp, modifyexp, endexp.
Each of these scripts now appends the showlog directly to the page,
instead of providing a link to a different page.

Also changed the showlog code to place the output into an inner iframe
so that it can be sized to fit in the window, without the outer stuff
(like the menus) getting scrolled away.