For the real NS parser, simulate the way raw object names work in the testbed parser.  For example, in:
  set node [$ns node]

The "set" creates a variable named node, but the testbed parser also
renames the object to be "node".  Doing a similar rename is hard in
the real NS parser since we have to change any existing references to
the object.  Instead, we dynamically create a procedure with the name
that forwards methods to the actual object.

And so, the corrupting of my soul continues.