Add the ability to include the local port number when reporting on sockets, so that we can correctly handle multiple sockets between two hosts.
The version of output used is controlled by the new
LIBNETMON_OUTPUTVERSION environment variable. The old output is
version 1, which is the default for now. The new version is 2.

Version 2 is still in ASCII. The address format has chaged from
'remoteip.remoteport' to 'localport:remoteip:remoteport'.

Version 2 will soon get the ability include information about when
sockets are created and destroyed.