Add support for announcing the closing and opening of sockets to the monitor. (output version 2)
Wrap read(), recv(), and recvmsg() so that we can detect EOF on a
socket by a zero-length read.

Run a function to announce the closing of sockets at exit time, in
case any are still open.

Wrap accept() so that we can deal with traffic sent (NOT recieved) by
servers as well.

Fix minor bug where I was forgetting to ntohs() a port number.