Per Jay's request, log plab isalive events originating from the plab service slivers in order to track elab<->node connectivity.
A timeval struct is appended to a file (one per node) inside
/usr/testbed/log/plabisalive.  The reporting rate has been cranked up to
once per minute.  Therefore, if all 650 plab nodes where reporting in, that
amounts to about 10 reports/sec, or 14KB/min leading to a disk consumption
rate of 21MB/day (617MB/mo or 7.5GB/year).  Not sure how long we are going
to be running this.