Created a new "Built With Emulab" logo. Actually, I spent the time to redraw the ping-pong balls in openoffice so we have an easily changeable figure; I got tired of pushing bits around with Photoshop.
The logo.sxd is just a big version of the ping pong balls. I then took
that and shrank it down to size to create builtwith.sxd, and added the
"Built with Emulab" text. Still didn't look quite right cause of
aliasing. So, I found the easiest thing to do was export it as a .eps
file, and then use ImageMagick's "convert" utility to create a .png
file:

	convert -antialias builtwith.eps builtwith.png

The antialiasing option makes it look much much nicer.

I have also changed the page footer so that when $TBMAINSITE=0, the
new logo is added to bottom of every page.

See http://www.emulab.net/builtwith.png