Plab connectivity monitoring changes:
* tmcd/tmcd.c

Updated connectivity log file names to include the date, and hence automatically
roll every 24 hours - much more friendly to the filesave, and easier to archive
off.

* tmcd/common/watchdog

Undo the conversion to minutes for the rusage time period (back to seconds)
to get the finer granularity Jay wants.  Not deployed yet.  Note that this
means the watchdog/rusage field will specify the number of _seconds_ not
minutes as it currently does.  I will update it's description to reflect this.