More work on the framework for running the monitor and stub automatically.
Now automatically determines the number of peers in the experiment.

run-monitor-libnetmon.sh automatically generates the IP mapping file
now.

Auto-create /local/logs if it doesn't exist (it doesn't on our
devbox image)

Add two new 'auto' scripts - these are intended to be run as part
of the automatic experiments. They properly wait for the stubs
to start before starting the monitors, etc. They log the outputs
of the stub and monitor to /local/logs, where loghole can pick
them up. They also kill their children if they are killed (to be
used with the program agent)

The auto-* scripts need a little more work in term of properly waiting
for their children to exit.