Added new utility, iperfd - intended to be a replacement for the server side of iperf. The main goal is to use much less memory than 'iperf -s'.
Supports many of the same same command line args, so that it can be
used as a dropin replacement for iperf.

Currently only uses about 40k of memory, regardless of how many
clients are connected.

Seems to behave exactly the same as an iperf server, but more testing
is needed.