Change the email going to testbed-errors from:
  From: user
  To: user
  Cc: testbed-errors

to

  From: testbed-ops
  To: user
  Bcc: testbed-errors
  X-NetBed-Cc: testbed-errors

This should cause all replies to these message to go to testbed-ops
instead of testbed-errors.  The only thing is that you need to be
careful when replying since if you only reply to the sender than it
will go to testbed-ops only and not to the user.  I was also thinking
about changing the other swap* failures messages still going to
testbed-ops in a similar way but due to the reply issue for those on
testbed-ops I will hold off on that for now unless someone else thinks
it's a good idea.

The addition of the X-NetBed-Cc header is to make filtering the email
easier since the fact that it is going to testbed-errors will no
longer be in the header.  This header is also present in the swap*
failures messages still going to testbed-ops.