New banner - same logo as the old banner, but now says 'emulab' instead of 'netbed', and has new text at the bottom.  We may do a redisgn of the page header, but for now, I was getting sick of the banner saying 'netbed'.
Approved by Jay.

The .svg file is the source used to generate the gifs. To generate
a gif, export the page from inkscape into a png, then use
'convert overlay.png overlay.gif' - we need a gif, rather than a
png, because IE's png transparency support is broken. The font used,
which you'll need to edit it, is Emblem, which is free.