Add a few features that Jon and I want to the modifyexp page:
Add a 'Syntax Check' button like the one on beginexp so that you
can syntax check without loosing all your edits if you had
syntax errors.

Add a file upload box and an option to use an NS file on ops,
like the beginexp page has.

Maybe we should try to find a spiffy ajax editor for this page?