More experiment run stuff. Also the NS file parsing support that was needed (reparsing of NS file with current bindings).
Mostly I decided that I made a huge mistake in using the exptidx as the
main cross index. I started to change that, which took a bunch of time.
Still more tables to fix up, but I think all of this code is gonna need
a major reorg once I am no longer under the gun to get it done.