Change to make mysql 5.X happy. The gist of this change is this that there is a query (by Mike I think) that has this form:
          SELECT * FROM t1, t2 JOIN t3 ON  ...

Prior to 5.X, the comma operator had the same precedence as join, and
so it was evaluated left to right. Well, now comma is lower then join
and this query breaks cause the right side is evaluated before the
left side. The solution is simply:

          SELECT * FROM (t1, t2) JOIN t3 ON  ...

but as the manual points out, why not just use:

          SELECT * FROM t1 JOIN t2 JOIN t3 ON ...

and avoid precedence issues.