Change the handling for when mysqld goes away (CR_SERVER_LOST || CR_SERVER_GONE_ERROR). Instead of bailing, sit and loop trying to reconnect, given that this is known to be a transient error, and we do not really want our daemons to go belly up during that brief time when the watchdog is getting it restarted. The query is then resent.
For the perl version of this change, I was a lot more pedantic since
we use this library from a zillion places. Also, there is some special
handling cause of the mysqld watchdog which would become useless if
the test query hung trying to reconnect to the server forever.

As a side effect of this change, we should see way less email when
mysqld goes catanoic since the new code will just loop instead of
generating tons of errors.

Might actually increase overall rebustness. On the other hand, could
end up being a total disaster!