Add an export option to the record listing. A new button on the Template Record page lets you export the contents of the archive that corresponds to that record, along with an XML file that describes the various DB bits for the template and instance.
This is just a first cut so that Mike can start playing around. Subject to
change, I'm sure.

The archive is dumped to /proj/$pid/exports/$guid/$vers/$exptidx, which
is basically the last commit of the instance when it was terminated.

The xml file is called export.xml and is placed in the top level directory
of the above directory. The file is created with XML::Simple, and a typical
XML file might look like:

<instance>
  <bindings>
    <name>NodeCount</name>
    <description>Number of nodes!</description>
    <value>1</value>
  </bindings>
  <bindings>
    <name>OS</name>
    <description></description>
    <value>RHL90-STD</value>
  </bindings>
  <bindings>
    <name>ScriptArgs</name>
    <description></description>
    <value>-b</value>
  </bindings>
  <eid>NewOne-V2</eid>
  <guid>10149/2</guid>
  <metadata>
    <name>M1</name>
    <guid>10162/1</guid>
    <value>Some metadata</value>
  </metadata>
  <pid>testbed</pid>
  <runs>
    <name>1</name>
    <archive_tag>T20060526-082533-172_endexp</archive_tag>
    <description></description>
    <exptidx>110</exptidx>
    <idx>1</idx>
    <runid>NewOne-V2</runid>
    <start_time>2006-05-26 08:23:02</start_time>
    <stop_time>2006-05-26 08:25:16</stop_time>
  </runs>
  <uid>stoller</uid>
</instance>