Enough already Mike!  A couple of changes for high latency (> 2 sec) links.
Change the rude.cfg script we generate to have a final "wait around" period
(MODIFY with 0 rate) before exiting.  We have it wait for the one-way latency
of the link.  This way, rude will not exit--and linktest will not do the
barrier synch--til all packets have had a fair chance to get to the other side.
This way we don't need a randomly chosen sleep after the barrier returns
and before we read the crude results.  Though rude already supported the
MODIFY with zero rate, I had to fix it since it didn't work if it was the
last command for the flow.

Adjust the timeout period for ping when doing the latency tests.  Previously
we waited for 3 seconds, now wait for max(RTT+1, 3) seconds.

Much more work will have to be done to support multi-second latencies
all around, this was just the low-hanging fruit.