Support for setting up GNU Radio on the node-side of things.
This operates pretty much the same way that wifi setup does, except that
the interface does not exist when rc.ifconfig runs.  Instead, it is
created as part of the generated rc.ifc script.  This required a couple of
little (non-invasive) hacks to libsetup.pm and rc.ifconfig.

Note that we are not passing gain and carrier sense threshold
parameters through to the radio interface setup, but they are easy to add.