Finish up "support" for ssh-mime (ssh to node) on Mac OSX.
Rather then use Tim's code (which requires that I learn about creating
packages on the Mac), I hacked up a little AppleScript widget. Sorta
goes like this:

* Copy ssh-mime.pl to your ~/bin directory on your Mac.

* mkdir ~/Library/Scripts/Folder\ Action\ Scripts

* Copy ssh-mime.scpt to the above directory.

* Start up Firefox and go into the Preferences->Downloads menu to figure
  out where files are being saved (typically ~/Desktop).

* Go into the Finder and find that folder. Right Click on that folder.
  (If you do not have a 3 button external mouse for your Mac like I do,
  then you select the folder, and do Control-Click).

* You will see an option to "Attach Folder Action" in the menu. Select
  that option.

* A finder file chooser pops up. Go to your

        ~/Library/Scripts/Folder\ Action\ Scripts

  and choose the ssh-mime.scpt entry.

Okay, so how does this work? The AppleScript ssh-mime.scpt file is now going
to be run whenever a file is downloaded (copied) into your firefox downloads
directory. The script will look at the name of the file and if it ends
in .tbssh, it will invoke ssh-mime.pl on it. Yep, its a hack, but its a lot
easier then learning the Xcode development environment!