Fix bug in libosload when there is an image problem.  The code that checks if there is an image problem is part of the code that looks up information on the image.  The problem is that this info is catched so that the next time this info is looked up the checks are not performed.  Thus, libosload will attempt to load the image even though there is a problem with the image.  This will often lead to obscure error messages.  Fixed by using a special value for when there is a problem with the image.
Fix bug when an experiment has more than 1 image type to load.  Only the
results from one of the image loads will be properly recognized. The
others will be ignored.  This is because the the same file handle is being
used for all the children due to the use of a global file handle glob.
Fixed by changing the glob to a variable and using "new IO::Handle" to
create a new handle for each child.