Bug fixes: Fix range check - this was necessary because of my change to the ackFor calculation.
Tricky detail about STL list.erase() - it does *not* erase the end iterator,
which meant that we were failing to remove the packet being acked (we were
removing all packets up to that one). So, we have it increment the iterator
in localAck() even if we find the one we're looking for.

Tested well for lossless connections - still needs testing for lossy
connections and SACK (see below). Hopefully, it works for the former,
but I know it doesn't work for the latter.

Added more debugging output to localAck()

Added a check which should warn us if we see any SACKS, though I
can't be sure it works, because I haven't seen any yet (that I know
of! :)