Fixed the traffic model so that it is now in line with what Rob and I discussed. Delays are before the write with the write size cached. Writes expire based on an expiration date.
Miscellaneous fixes. Open problem: PacketSensor does not deal correctly with zero-sized packets. I changed KernelTcp to pass such packets because otherwise there is no way to do state changes based on SYN/FIN/other packets. SYN is handled ok for now because of a change noted below. FIN is not.

Added a StateSensor. As I discovered, using the kernel tcp_info data structure isn't useful when dealing with fields that change on a packet by packet basis because the kernel information is retrieved at processing time and not capture time. For instance, it is useless when trying to determine whether a connection was established by the time a particular packet was sent (to determine whether it is part of the three-way-handshake). The StateSensor keeps track of the state machine and correlates it to the packet involved. This allows the other sensors to rely on it to distinguish between connection setup and the rest of the connection traffic. Added references to it in the PacketSensor, the DelaySensor and the ThroughputSensor.

Changed the way packet information was transmitted to the sensors to make it easier to add new packet types (as will be necessary when accept()s are handled).

Fixed all outstanding flaws in the basic feedback mechanism. In short, "Its alive!". Currently the only data being transmitted is the base rtt (MinDelay). Now that the feedback is working in a basic form, it will be easier to get the other characteristics online.