Fix a few bugs in argument handling for replay.
First, we can't use logWrite() in processArgs(), because it has not
been initialized yet. (And, presumably, the log writer *has* to be
initialized later, because at some point we will want to add command
line args to control its behavior).

Second, actually exit if there are any errors, instead of proceeding
to do the wrong thin.