Add patch to modify Mysql.pm to allow setting the "InactiveDestroy" in the underlying DB handle.  Also avoid disconnecting the file handle explistly on DESTROY as that will be taken care of in the DESTROY method for the the DB handle.
Override perl version of fork() to set InactiveDestroy in all open
database handles in the child so that it won't send a disconnect when
the handle is destroyed as this will also close the database handle
for the parent.  It will also call tblog_new_child_process in the
child process to properly inform tblog of the new process. This will
be a NoOp if the libtblog module is not loaded.