Okay, this is a nasty little hack ... Add support for a global delays reset. I've done this with an event group cause otherwise I was going to get sucked into the event system and spit out the other end. You can reset the delays in your experiment either from the ns file:
	$ns at 100 "$ns reset-lans"

or from the command line:

	tevc -e foo/bar now all_lans reset

and yes, "all_lans" is a magic token.

It would be nice to support per-link or lan reset, but that is going
to require reorganizing the delay start up scripts on the delay nodes,
since right now a single delay agent operates for muliple links and lans.