plab logging enhancements.
timing information for various RPCs is now logged to
/usr/testbed/log/plabtiming.log.  This info will be useful for extracting
trends for the various plab nodes, and in calculating reliability and
timing metrics.  These could be used, for e.g., to pick nodes that tend to
come up more quickly.

This update also squelches much of the python backtrace noise when plab nodes
fail to setup correctly (can be turned on with debug flag).  Instead, failures
are summarized on a single line.

Oh, and pay no attention to the aspect behind the curtain!  Yes, you may
groan and moan if you wish - I'm using aspects to help do the logging.  I
find this to be a really slick way of wrapping several functions!